# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IT Architect Assistant Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/knowledge/bases": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Knowledge Bases",
        "operationId": "list_knowledge_bases_api_v1_knowledge_bases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeBaseResponse"
                  },
                  "type": "array",
                  "title": "Response List Knowledge Bases Api V1 Knowledge Bases Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Create Knowledge Base",
        "operationId": "create_knowledge_base_api_v1_knowledge_bases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/bases/{knowledge_base_id}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Knowledge Base",
        "operationId": "get_knowledge_base_api_v1_knowledge_bases__knowledge_base_id__get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "knowledge"
        ],
        "summary": "Update Knowledge Base",
        "operationId": "update_knowledge_base_api_v1_knowledge_bases__knowledge_base_id__patch",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/bases/{knowledge_base_id}/documents": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Base Documents",
        "operationId": "list_base_documents_api_v1_knowledge_bases__knowledge_base_id__documents_get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "knowledge_version_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge Version Id"
            }
          },
          {
            "name": "include_deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Deleted"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeBaseDocumentResponse"
                  },
                  "title": "Response List Base Documents Api V1 Knowledge Bases  Knowledge Base Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/bases/{knowledge_base_id}/update-runs": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Base Update Runs",
        "operationId": "list_base_update_runs_api_v1_knowledge_bases__knowledge_base_id__update_runs_get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeUpdateRunResponse"
                  },
                  "title": "Response List Base Update Runs Api V1 Knowledge Bases  Knowledge Base Id  Update Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/bases/{knowledge_base_id}/sync": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Start Base Sync",
        "operationId": "start_base_sync_api_v1_knowledge_bases__knowledge_base_id__sync_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "execute_inline",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Execute Inline"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeUpdateRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/bases/{knowledge_base_id}/reindex": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Reindex Knowledge Base",
        "operationId": "reindex_knowledge_base_api_v1_knowledge_bases__knowledge_base_id__reindex_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeReindexRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeUpdateRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/bases/{knowledge_base_id}/embedding-profile": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Switch Knowledge Base Embedding Profile",
        "operationId": "switch_knowledge_base_embedding_profile_api_v1_knowledge_bases__knowledge_base_id__embedding_profile_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeEmbeddingProfileSwitchRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeUpdateRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/bases/{knowledge_base_id}/select": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Select Knowledge Base",
        "operationId": "select_knowledge_base_api_v1_knowledge_bases__knowledge_base_id__select_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseSelectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/bundles/import": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Import Bundle",
        "operationId": "import_bundle_api_v1_knowledge_bundles_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBundleImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBundleImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeBaseCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "KnowledgeBaseCreateRequest"
      },
...
      "KnowledgeBaseKind": {
        "type": "string",
        "enum": [
          "system_mandatory",
          "user_managed"
        ],
        "title": "KnowledgeBaseKind"
      },
      "KnowledgeBaseResponse": {
        "properties": {
          "knowledge_base_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "title": "Knowledge Base Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "kind": {
            "$ref": "#/components/schemas/KnowledgeBaseKind"
          },
          "status": {
            "$ref": "#/components/schemas/KnowledgeBaseStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "active_knowledge_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Knowledge Version Id"
          },
          "active_version_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version No"
          },
          "selected_for_generation": {
            "type": "boolean",
            "title": "Selected For Generation",
            "default": false
          },
          "selected_knowledge_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Knowledge Version Id"
          },
          "selected_knowledge_version_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Knowledge Version No"
          },
          "active_embedding_space_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Embedding Space Id"
          },
          "active_embedding_space_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Embedding Space Code"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count",
            "default": 0
          },
          "active_source_count": {
            "type": "integer",
            "title": "Active Source Count",
            "default": 0
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          },
          "latest_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Sync At"
          },
          "latest_sync_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeUpdateStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_sync_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Sync Run Id"
          },
          "latest_successful_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Successful Sync At"
          },
          "last_sync_duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync Duration Sec"
          },
          "last_sync_error_count": {
            "type": "integer",
            "title": "Last Sync Error Count",
            "default": 0
          },
          "versions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "knowledge_base_id",
          "code",
          "name",
          "kind",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeBaseResponse"
      },
      "KnowledgeBaseSelectRequest": {
        "properties": {
          "knowledge_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Version Id"
          }
        },
        "type": "object",
        "title": "KnowledgeBaseSelectRequest"
      },
...
      "KnowledgeBaseUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeBaseStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "KnowledgeBaseUpdateRequest"
      },
      "KnowledgeBundleImportRequest": {
        "properties": {
          "manifest_uri": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Manifest Uri"
          },
          "knowledge_base_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Base Id"
          },
          "activate_if_validated": {
            "type": "boolean",
            "title": "Activate If Validated",
            "default": false
          },
          "execute_update_inline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execute Update Inline"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          }
        },
        "type": "object",
        "required": [
          "manifest_uri"
        ],
        "title": "KnowledgeBundleImportRequest"
      },
...
      "KnowledgeEmbeddingProfileSwitchRequest": {
        "properties": {
          "target_embedding_profile": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Target Embedding Profile"
          },
          "execute_inline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execute Inline"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "target_embedding_profile"
        ],
        "title": "KnowledgeEmbeddingProfileSwitchRequest"
      },
...
      "KnowledgeReindexRequest": {
        "properties": {
          "execute_inline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execute Inline"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "KnowledgeReindexRequest"
      },
...
    }
  }
}